add colors/colors-safe.d.ts
